export standalone codec (encoding and decoding functions)
rename `create*` to `encode*` and `parse*` to `decode*`
